refactor: response 빌더로 변환
